♻️ refactor: expose internal modules for testing purposes
This commit introduces a new section that exposes several internal modules (Config, Database, UI, ComponentFactory, PageHandlers, Router) to the `window.qwiklabs_testing` object.

This exposure is conditional, only occurring when `process` is defined and `process.env.NODE_ENV` is set to 'test'. This change is made to facilitate easier unit and integration testing of these modules in a test environment without affecting the production build.